refactor: update conformance standards and modernize runtime library
This commit significantly expands the repository's coding standards and
applies them across the runtime library and its test suite.

- Updated `check-code-conformance` skill with a detailed automation
  coverage map for `clang-tidy` and a residual manual review checklist.
- Modernized runtime code and tests using C++20/23 features such as
  `std::numbers::e`, `[[maybe_unused]]`, and `std::array`.
- Enhanced type safety and consistency in `TrackListProjection` and
  related components (e.g., explicit `ViewId` casts).
- Standardized member access and naming conventions across several
  runtime classes.
- Applied `final` specifiers to concrete structs and classes in
  accordance with the updated style guide.
- Expanded `doc/design/type-to-header-map.md` with comprehensive mappings
  for Gtk, Gio, Gdk, and Gsk types.

+++ app/runtime/AllTracksSource.h
@@ -9,7 +9,6 @@
 #include <ao/lmdb/Transaction.h>
 
 #include <flat_set>
-#include <vector>
 
 namespace ao::rt
 {
@@ -35,7 +34,7 @@ namespace ao::rt
 
     // TrackSource interface
     std::size_t size() const override { return _trackIds.size(); }
-    TrackId trackIdAt(std::size_t index) const override { return *(_trackIds.begin() + index); }
+    TrackId trackIdAt(std::size_t index) const override { return *(_trackIds.begin() + static_cast<std::ptrdiff_t>(index)); }
     std::optional<std::size_t> indexOf(TrackId id) const override;
 
   private:
