+++ include/ao/library/Meta.h
@@ -13,7 +13,7 @@ namespace ao::library
   constexpr std::uint32_t kLibraryMetaMagic = 0x42534C52U;
   constexpr std::uint32_t kLibraryVersion = 1;
 
-  enum class MetaRecordId : std::uint32_t
+  enum class MetaRecordId : std::uint8_t
   {
     Header = 1,
   };
@@ -28,7 +28,7 @@ namespace ao::library
     std::array<std::byte, 16> libraryId;
   };
 
-  static_assert(sizeof(MetaHeader) == 48, "MetaHeader must be exactly 48 bytes");
+  static_assert(sizeof(MetaHeader) == 48, "MetaHeader must be exactly 48 bytes"); // NOLINT(readability-magic-numbers)
   static_assert(std::is_trivially_copyable_v<MetaHeader>, "MetaHeader must be trivially copyable");
   static_assert(std::is_standard_layout_v<MetaHeader>, "MetaHeader must have standard layout");
 }
